Extract article rendering into src/article.rs (#47)
Moved try_fetch_readme, markdown_to_styled_lines,
fetch_and_extract_article, extract_article_content,
annotations_to_style, html_to_styled_lines, and MAX_RESPONSE_BYTES
out of app.rs into a new article module. app.rs drops from 1168 to
785 lines and no longer imports html2text/ratatui styling or the
theme module directly.

Only fetch_and_extract_article and html_to_styled_lines are pub — the
rest are private helpers of the new module.
