+++ src/main.rs
@@ -1,5 +1,6 @@
 mod api;
 mod app;
+mod article;
 mod event;
 mod keys;
 mod state;
